Separated ui from business logic
