+++ ruby-2/project/main.rb
@@ -1,0 +1,5 @@
+#encoding: utf-8
+
+require_relative 'forca'
+
+jogo_da_forca
